Change error beahvior to fit gettext() behavior
Although it might be a good idea to use exceptions, we use null to
indicate errors as we have to fail silently, as this is ext/gettext
behavior. Yes we might just catch every exception, but this doesn't
seem to be very intelligent at all, if we later change the code
and a real exception might be raised somewhere.

Signed-off-by: David Soria Parra <sn_@gmx.net>